Move InputLike implementations into their own files, add ScanCode impl
