+++ src/scan_codes/wasm.rs
@@ -1,9 +1,15 @@
 //! Helper enum to define scan codes on the QWERTY keyboard layout.
 
+use crate::input_like::ReflectInputLike;
+use bevy::reflect::Reflect;
+use serde::{Deserialize, Serialize};
+
 /// The key locations as defined by the keys on the QWERTY keyboard layout.
 ///
 /// The [`u32`] representation of this enum are the `KeyboardEvent.code` values on Wasm.
 /// See <https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/keyCode#value_of_keycode>.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize, Reflect)]
+#[reflect(InputLike)]
 #[repr(u32)]
 pub enum QwertyScanCode {
     /// The location of the `1` key on the QWERTY keyboard layout.
